requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
google-generativeai
pydantic
python-multipart
